requirements.txt
attrs==21.2.0
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
coverage==6.2
dnspython==2.1.0
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
iniconfig==1.1.1
mypy-extensions==0.4.3
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
protobuf==3.19.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.12.1
pyparsing==3.0.6
pytest==6.2.5
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
